+++ src/REvoDesign/tools/pymol_utils.py
@@ -571,3 +571,7 @@ def any_posision_has_been_selected():
         if x == "sele"
     ]
     return bool(selected_positions)
+
+
+def get_all_groups(enabled_only: bool = False) -> List[str]:
+    return cmd.get_names("group_objects", int(enabled_only))